Make ConnectionFactory#networkRecoveryInterval a long
There is no real need for it to be a long but other libraries and JVM languages
may use longs for various reasons and using a long with an int field is

 * Unsafe
 * Not entirely obvious to some developers

+++ test/src/com/rabbitmq/client/test/functional/ConnectionRecovery.java
@@ -18,7 +18,7 @@ import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class ConnectionRecovery extends BrokerTestCase {
-    public static final int RECOVERY_INTERVAL = 2000;
+    public static final long RECOVERY_INTERVAL = 2000;
 
     public void testConnectionRecovery() throws IOException, InterruptedException {
         assertTrue(connection.isOpen());
